bazel: require cgroup sandboxing in CI
This is required so we can run tests that use the local network in parallel. We can check it's working in our tools/bazel script, which runs as a hook before bazel is invoked.

From the bazel documentation (we want linux sandboxing enabled):

```
linux-sandbox goes one step further and builds on top of the processwrapper-sandbox. Similar to what Docker does under the hood, it uses Linux Namespaces (User, Mount, PID, Network and IPC namespaces) to isolate the action from the host. That is, it makes the entire filesystem read-only except for the sandbox directory, so the action cannot accidentally modify anything on the host filesystem. This prevents situations like a buggy test accidentally rm -rf'ing your $HOME directory. Optionally, you can also prevent the action from accessing the network. linux-sandbox uses PID namespaces to prevent the action from seeing any other processes and to reliably kill all processes (even daemons spawned by the action) at the end.

Both the linux-sandbox and the darwin-sandbox do not work in a "nested" scenario due to restrictions in the mechanisms provided by the operating systems. Because Docker also uses Linux namespaces for its container magic, you cannot easily run linux-sandbox inside a Docker container, unless you use docker run --privileged. On macOS, you cannot run sandbox-exec inside a process that's already being sandboxed. Thus, in these cases, Bazel automatically falls back to using processwrapper-sandbox.
```

We want to not implicitly fallback, but some tests require no sandboxing (ie. //:gazelle_test). Because we run in a docker container for CI, we want to ensure sandboxing is actually enabled.

The check is done in the same way Bazel itself checks if the sandbox is functioning, by ensuring the true command exits successfully in the sandbox (the daemon caches this result).